pyproject.toml
[project]
name = "korapay-client"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Gbenga Adeyi", email = "adeyigbenga005@gmail.com" }
]
dependencies = [
    "httpx>=0.27.0",
    "pycryptodome>=3.20.0",
    "pydantic[email]>=2.6.4",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.3.4",
    "python-dotenv>=1.0.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/korapay_client"]
